+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "0.0.1a"
 authors = [{ name = "Jan Roth", email = "r0th.j4n@gmail.com" }]
 description = "Mimetic Finite Element Solver for 2D."
 requires-python = ">=3.10"
-dependencies = ["numpy>=2.0.0", "scipy", "jax", "pyvista"]
+dependencies = ["numpy>=2.0.0", "scipy", "pyvista"]
 
 
 [build-system]
@@ -55,6 +55,8 @@ docs = [
     "sphinxcontrib-bibtex",
     "sphinx_book_theme",
     "jupyter-sphinx",
+    "imageio",
+    "rmsh@git+https://github.com/j4nr0th/rmesh",
 ]
